Hoofdstuk 4 Opdracht 5
This is my attempt for Task 5 of Chapter 4. I decided to try and get the value to match a full day. That means that if you multiply the value for Seconds by a set amount of days. Say 14. Then you get to see the value two weeks is worth in the format of seconds hours months and years. I also decided to put the value of every potential kind of month in there for those who specifically want to see the value for February or a month that's 30 days instead of 31.