Hoofdstuk 6
This chapter is still very incomplete because it has a whopping NINE tasks.

However, it's not as grand as it first appears to be. Tasks 7-9 use the example made in chapter 6.5, this means they're basically copies of each others with minor changes. I believe the rest of the tasks are unique though, but they still reference each other sometimes. Basically, just take it one at a time. I should be done within this week easily.