Hoofdstuk 6 1st Set
This is the 1st set of tasks for chapter 6 that I finished. Some of them have extra versions because I wanted to play around with what I could do with the previous tasks. Task 6 was also going to have this, but it ended up being too time consuming, so I moved on.