Hoofdstuk 7 (Zonder Verdiepings Opdrachten)
Just like chapter 6, and i think 5, I haven't done the extra tasks for this yet. However, I still completed the chapter so I'm pushing this to my Repository. I'll try to revisit some of this stuff later this week to still see if I understand it. If not this week, I'll be doing it before the final chapter at the very least. This way I can be sure that I remembered the material later in the future.